[pkg] move u1db url to official requirements file
u1db was declared on requirements.pip as 'u1db' and as a url on
requirements-latest.pip; That was ok until they were the same, but now
we depend on the exact one provided by this url as the newest one has a
couple of errors.
setup.py was also adapted to accept url for develop mode.